#79 refactor: 컴포넌트 및 파일 구조 개선 - 파일명 PascalCase 컨벤션으로 통일 (ChartHeader, MonthlyChartTabs 등) - MonthlyChartContent, MonthlyChartList 등 의미가 명확하게 전달되도록 이름 수정 - 불명확했던 IdolChart → MonthlyChartSection으로 리네이밍 - 전체 구조 가독성 및 유지보수성 향상 목적
