+++ src/pages/main-page/components/ChartHeader.jsx
@@ -1,7 +1,7 @@
 import Button from '@components/common/Button';
-import Chart from '@assets/icons/icon_chart';
+import IconChart from '@assets/icons/icon_chart';
 
-const ChartHeader = () => {
+const MonthlyChartHeader = () => {
   return (
     <div className='mt-4 mb-8 flex items-baseline justify-between md:max-w-5/8'>
       <h1 className='title-text text-white'>이달의 차트</h1>
@@ -14,11 +14,11 @@ const ChartHeader = () => {
           console.log('모달');
         }}
       >
-        <Chart />
+        <IconChart />
         <p className='ml-1'>차트 투표하기</p>
       </Button>
     </div>
   );
 };
 
-export default ChartHeader;
+export default MonthlyChartHeader;
