+++ src/pages/main-page/components/MonthlyChartList.jsx
@@ -7,7 +7,7 @@
  * @returns {JSX.Element}
  */
 
-const IdolListTabContent = ({ idolData, IdolList }) => {
+const MonthlyChartList = ({ idolData, IdolList }) => {
   //idolData가 제대로 불러오지 않았을 경우 처리
   if (!idolData || idolData.length === 0) {
     return <div>아이돌 정보가 없습니다.</div>;
@@ -36,4 +36,4 @@ const IdolListTabContent = ({ idolData, IdolList }) => {
   );
 };
 
-export default IdolListTabContent;
+export default MonthlyChartList;
