--- src/pages/main-page/sections/IdolChart.jsx
@@ -1,13 +1,0 @@
-import ChartHeader from '../components/ChartHeader';
-import IdolListTab from './components/Idol-list-tab';
-
-const IdolChart = () => {
-  return (
-    <>
-      <ChartHeader />
-      <IdolListTab />
-    </>
-  );
-};
-
-export default IdolChart;
