+++ src/pages/main-page/sections/MonthlyChartSection.jsx
@@ -1,0 +1,13 @@
+import MonthlyChartHeader from '../components/ChartHeader';
+import MonthlyChartTabs from '../components/MonthlyChartTabs';
+
+const MonthlyChartSection = () => {
+  return (
+    <>
+      <MonthlyChartHeader />
+      <MonthlyChartTabs />
+    </>
+  );
+};
+
+export default MonthlyChartSection;
